+++ CHANGELOG.md
@@ -83,6 +83,12 @@
 
 
 #### BUG FIXES:
+* Appsec
+  * Fixed plugin crash if `exceptions` block is passed as empty in `akamai_appsec_siem_settings` resource
+
+
+
+
 
 * Cloud Access
   * Resolved issues with drift detection after deletion of key version in `akamai_cloudaccess_key` resource. ([I#579](https://github.com/akamai/terraform-provider-akamai/issues/579))